feat: moved unlock state to wallet
